refactor: rename to shell to make more generic
